refactor: start small code cleanup
