+++ sgxvm/src/types.rs
@@ -11,6 +11,7 @@ use std::{
     string::String,
 };
 use std::boxed::Box;
+use sgx_types::*;
 
 pub static GASOMETER_CONFIG: Config = Config::london();
 
@@ -91,4 +92,27 @@ impl ExecutionResult {
             data,
         }
     }
+}
+
+#[repr(C)]
+pub struct AllocationWithResult {
+    pub result_ptr: *mut u8,
+    pub result_len: usize,
+    pub status: sgx_status_t
+}
+
+impl Default for AllocationWithResult {
+    fn default() -> Self {
+        AllocationWithResult {
+            result_ptr: std::ptr::null_mut(),
+            result_len: 0,
+            status: sgx_status_t::SGX_ERROR_UNEXPECTED,
+        }
+    }
+}
+
+#[repr(C)]
+pub struct Allocation {
+    pub result_ptr: *mut u8,
+    pub result_size: usize,
 }
